refactor: merge channels_v2 exceptions into channels/exceptions.py
