+++ apps/channels/tests/channels/stages/test_consent_check.py
@@ -3,8 +3,8 @@ from unittest.mock import MagicMock
 import pytest
 
 from apps.channels.channels_v2.capabilities import PlatformConsentConfig
-from apps.channels.channels_v2.exceptions import EarlyAbort
 from apps.channels.channels_v2.stages.core import ConsentCheckStage
+from apps.channels.exceptions import EarlyAbort
 from apps.channels.models import ChannelPlatform
 from apps.channels.tests.channels.conftest import make_capabilities, make_context
 from apps.experiments.models import ParticipantData
